# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task API
  description: A simple task management API built with Express and Sequelize
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /api/task:
    get:
      summary: Get all tasks
      description: Retrieves all tasks ordered by order_index
      operationId: getTasks
      tags:
        - Tasks
      responses:
        "200":
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create a new task
      description: Creates a new task with the provided description and inserts it at the bottom of the list
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
              properties:
                description:
                  type: string
                  description: The task description
                  example: Buy groceries
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad request - description is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/task/complete:
    post:
      summary: Mark a task as complete
      description: Sets the date_completed field to the current timestamp
      operationId: completeTask
      tags:
        - Tasks
      parameters:
        - name: id
          in: query
          required: true
          description: The ID of the task to complete
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Task marked as complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad request - task ID is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/task/delete:
    delete:
      summary: Delete a task
      description: Permanently removes a task from the database
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - name: id
          in: query
          required: true
          description: The ID of the task to delete
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Task deleted successfully
        "400":
          description: Bad request - task ID is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/task/order:
    post:
      summary: Reorder tasks
      description: Updates the order_index of tasks based on the provided array of IDs
      operationId: reorderTasks
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - order
              properties:
                order:
                  type: array
                  items:
                    type: integer
                  description: Array of task IDs in the desired order
                  example: [1, 3, 2]
      responses:
        "200":
          description: Tasks reordered successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "400":
          description: Bad request - order list is required or invalid format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/task/completed:
    delete:
      summary: Delete all completed tasks
      description: Permanently removes all tasks that have a date_completed value set
      operationId: deleteCompletedTasks
      tags:
        - Tasks
      responses:
        "200":
          description: Completed tasks deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Deleted 5 completed tasks
                  count:
                    type: integer
                    description: Number of tasks deleted
                    example: 5
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/task/edit:
    post:
      summary: Edit a task description
      description: Finds a task by its current description and updates it to a new description
      operationId: editTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_task_description
                - new_task_description
              properties:
                old_task_description:
                  type: string
                  description: The current description of the task to find
                  example: Buy groceries
                new_task_description:
                  type: string
                  description: The new description to set
                  example: Buy organic groceries
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad request - both descriptions are required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        date_created:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: "2024-01-15T10:30:00.000Z"
        date_completed:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the task was completed (null if not completed)
          example: null
        description:
          type: string
          description: The task description
          example: Buy groceries
        order_index:
          type: integer
          description: The order position of the task
          example: 1

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Something went wrong